Updated for DEV C++

+++ DAA_Labs/Lab3/HeapSort.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-long long steps = 0;
+long steps = 0;
 
 void heapify(int arr[], int n, int i) {
     steps++;
